+++ src/models/ComponentStyle.js
@@ -1,11 +1,9 @@
 // @flow
 import hashStr from 'glamor/lib/hash'
+import stylis from 'stylis'
 
 import type { RuleSet, NameGenerator } from '../types'
 import flatten from '../utils/flatten'
-import parse from '../vendor/postcss-safe-parser/parse'
-import postcssNested from '../vendor/postcss-nested'
-import autoprefix from '../utils/autoprefix'
 import styleSheet from './StyleSheet'
 
 /*
@@ -44,10 +42,8 @@ export default (nameGenerator: NameGenerator) => {
       if (!inserted[hash]) {
         const selector = nameGenerator(hash)
         inserted[hash] = selector
-        const root = parse(`.${selector} { ${flatCSS} }`)
-        postcssNested(root)
-        autoprefix(root)
-        this.insertedRule.appendRule(`\n${root.toResult().css}`)
+        const css = stylis(`.${selector}`, flatCSS, false, false)
+        this.insertedRule.appendRule(css)
       }
       return inserted[hash]
     }