+++ src/models/GlobalStyle.js
@@ -1,6 +1,5 @@
 // @flow
-import parse from '../vendor/postcss-safe-parser/parse'
-import postcssNested from '../vendor/postcss-nested'
+import stylis from 'stylis'
 
 import type { RuleSet } from '../types'
 import flatten from '../utils/flatten'
@@ -17,12 +16,9 @@ export default class ComponentStyle {
 
   generateAndInject() {
     if (!styleSheet.injected) styleSheet.inject()
-    let flatCSS = flatten(this.rules).join('')
-    if (this.selector) {
-      flatCSS = `${this.selector} {${flatCSS}\n}`
-    }
-    const root = parse(flatCSS)
-    postcssNested(root)
-    styleSheet.insert(root.toResult().css)
+    const flatCSS = flatten(this.rules).join('')
+    const cssString = this.selector ? `${this.selector} { ${flatCSS} }` : flatCSS
+    const css = stylis('', cssString, false, false)
+    styleSheet.insert(css)
   }
 }